Refactor attack style tracking to manager
