+++ runelite-client/src/main/java/net/runelite/client/events/AttackStyleChanged.java
@@ -1,0 +1,10 @@
+package net.runelite.client.events;
+
+import lombok.Value;
+import net.runelite.client.game.attackstyles.AttackStyle;
+
+@Value
+public class AttackStyleChanged
+{
+	private final AttackStyle attackStyle;
+}
