+++ runelite-client/src/main/java/net/runelite/client/game/attackstyles/WeaponType.java
@@ -1,10 +1,10 @@
-package net.runelite.client.plugins.xptracker;
+package net.runelite.client.game.attackstyles;
 
 import java.util.HashMap;
 import java.util.Map;
-import static net.runelite.client.plugins.xptracker.AttackStyle.*;
+import static net.runelite.client.game.attackstyles.AttackStyle.*;
 
-enum WeaponType
+public enum WeaponType
 {
 	TYPE_0(ACCURATE, AGGRESSIVE, null, DEFENSIVE),
 	TYPE_1(ACCURATE, AGGRESSIVE, AGGRESSIVE, DEFENSIVE),
